+++ apps/health_campaign_field_worker_app/lib/utils/constants.dart
@@ -170,6 +170,8 @@ class Constants {
     _version = version;
   }
 
+  static const String closedHouseholdSvg = 'assets/icons/svg/closed_household.svg';
+
   static List<RemoteRepository> getRemoteRepositories(
     Dio dio,
     Map<DataModelType, Map<ApiOperation, String>> actionMap,