+++ packages/closed_household/lib/pages/closed_household_details.dart
@@ -1,17 +1,14 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:closed_household/blocs/closed_household.dart';
 import 'package:closed_household/closed_household.dart';
 import 'package:digit_components/digit_components.dart';
 import 'package:digit_components/widgets/digit_sync_dialog.dart';
 import 'package:digit_components/widgets/atoms/text_block.dart';
 import 'package:digit_data_model/data_model.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import '../../utils/i18_key_constants.dart' as i18;
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:reactive_forms/reactive_forms.dart';
 
-import '../models/entities/user_action.dart';
+import '../../utils/i18_key_constants.dart' as i18;
 import '../router/closed_household_router.gm.dart';
 import '../widgets/back_navigation_help_header.dart';
 import '../widgets/localized.dart';