+++ packages/closed_household/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.0-dev.1
 
-* TODO: Describe initial release.
+* Initial release of closed household flow
+